+++ A-4E-C/Options/optionsDb.lua
@@ -22,13 +22,6 @@ local tblCPLocalList = oms.getTblCPLocalList(modulelocation)
 
 
 return {
-
-	rwrType					= DbOption.new():setValue(0):combo({DbOption.Item(_('AN/APR-23 (Sound only)')):Value(0),
-																DbOption.Item(_('AN/APR-25 (Display)')):Value(1),
-																--[[DbOption.Item(_('ON')):Value(2),
-																DbOption.Item(_('DISMOUNTED')):Value(3),]]--
-																}),
-
 	trimSpeed				= DbOption.new():setValue(0):combo({DbOption.Item(_('100% (default)')):Value(0),
 																DbOption.Item(_('50%')):Value(1),																
 																}),